:art: refactoring - Comment, Question 내부의 uid 타입 변경 (int => Long)
- findUidByEmailRepo 반환 타입 변경 (Integer => Long)
- 해시맵 저장시 오류가 발생하는거같은데 왜 -1이 저장되는지 모르겠음
- 프로시저 저장시 안돌아가는듯

+++ src/main/java/com/example/androidserver/Comment/mapper/CommentRowMapper.java
@@ -13,7 +13,7 @@ public class CommentRowMapper implements RowMapper<Comment> {
     public Comment mapRow(ResultSet rs, int rowNum) throws SQLException {
         Comment comment = new Comment();
         comment.setCid(rs.getInt("cid"));
-        comment.setUid(rs.getInt("uid"));
+        comment.setUid(rs.getLong("uid"));
         comment.setQid(rs.getInt("qid"));
         comment.setContent(rs.getString("content"));
         comment.setCreateAt(rs.getTimestamp("createAt"));
